refactor: 将 practice-lesson-popover 单独提出来
