+++ src/pages/lessonstable/_constants/split-name-room.ts
@@ -1,6 +1,6 @@
 const NAME_ROOM_RE = /^(\P{ASCII}*)([\s\S]*)$/u;
 
 export function splitNameAndRoom(str: string) {
-  const match = str.match(NAME_ROOM_RE);
+  const match = NAME_ROOM_RE.exec(str);
   return match ? [match[1], match[2]] : [str, ""];
 }
